Fixed silly warnings.

+++ test/calculator_test.cpp
@@ -48,7 +48,7 @@ protected:
     }
 
 protected:
-    static constexpr size_t kCalculatorTimeIntervalInMilliseconds = 60 * 1000;
+    static constexpr int64_t kCalculatorTimeIntervalInMilliseconds = 60 * 1000;
     ITimeIntervalSumCalculatorPtr<int64_t> m_calculator = std::make_shared<TimeIntervalSumCalculator<int64_t>>(
         kCalculatorTimeIntervalInMilliseconds, 1024, m_timeProvider
     );
@@ -99,13 +99,13 @@ TEST_F(TimeIntervalCalculatorTest, SumAfterLongTimeWithUniformInsertDistribution
 
 TEST_F(TimeIntervalCalculatorTest, SumDuringContiniusExpirityTest)
 {
-    for (size_t i = 0; i < kCalculatorTimeIntervalInMilliseconds; ++i)
+    for (int64_t i = 0; i < kCalculatorTimeIntervalInMilliseconds; ++i)
     {
         m_calculator->put(1);
         waitFor(1);
     }
 
-    for (size_t i = 1 ; i < kCalculatorTimeIntervalInMilliseconds; ++i)
+    for (int64_t i = 1 ; i < kCalculatorTimeIntervalInMilliseconds; ++i)
     {
         m_calculator->put(0);
         ASSERT_EQ(m_calculator->get(), kCalculatorTimeIntervalInMilliseconds - i);
@@ -115,13 +115,13 @@ TEST_F(TimeIntervalCalculatorTest, SumDuringContiniusExpirityTest)
 
 TEST_F(TimeIntervalCalculatorTest, SumWithNegativeNumbersTest)
 {
-    for (size_t i = 0; i < 100; ++i)
+    for (int64_t i = 0; i < 100; ++i)
     {
         m_calculator->put(1);
     }
     ASSERT_EQ(m_calculator->get(), 100);
 
-    for (size_t i = 0; i < 100; ++i)
+    for (int64_t i = 0; i < 100; ++i)
     {
         m_calculator->put(-1);
     }
@@ -130,12 +130,12 @@ TEST_F(TimeIntervalCalculatorTest, SumWithNegativeNumbersTest)
 
 TEST_F(TimeIntervalCalculatorTest, SumWithRealloctionTest)
 {
-    for (size_t i = 0; i < 10000; ++i)
+    for (int64_t i = 0; i < 10000; ++i)
     {
         m_calculator->put(1);
     }
     waitFor(2 * kCalculatorTimeIntervalInMilliseconds / 3);
-    for (size_t i = 0; i < 1000; ++i)
+    for (int64_t i = 0; i < 1000; ++i)
     {
         m_calculator->put(1);
     }
@@ -143,7 +143,7 @@ TEST_F(TimeIntervalCalculatorTest, SumWithRealloctionTest)
     m_calculator->put(0);
     ASSERT_EQ(m_calculator->get(), 1000);
 
-    for (size_t i = 0; i < 1000; ++i)
+    for (int64_t i = 0; i < 1000; ++i)
     {
         m_calculator->put(-1);
     }
